% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unifont.R
\name{unifont}
\alias{unifont}
\title{Load Unifont font}
\usage{
unifont(upper = TRUE, jp = FALSE, csur = TRUE, sample = FALSE)
}
\arguments{
\item{upper}{Include glyphs above the Unicode Basic Multilingual plane.}

\item{jp}{Use Japanese version of Chinese characters.}

\item{csur}{Include (Under-)Conscript Unicode Registry glyphs.}

\item{sample}{Add circle to "Combining" characters.}
}
\value{
A \code{\link[bittermelon:bm_font]{bittermelon::bm_font()}} object.
}
\description{
The function \code{unifont()} loads in several GNU Unifont hex files as a
single \code{\link[bittermelon:bm_font]{bittermelon::bm_font()}} object.
}
\examples{
if (require("bittermelon")) {
  # Note Unifont is a **big** font
  system.time(font <- unifont())
  length(font)
  object.size(font)

  # Mandarin Chinese
  bml <- as_bm_list("\uff32\u5f88\u68d2\uff01", font = font)
  bm <- do.call(cbind, bml)
  print(bm, px = px_ascii)

  # Emoji
  bml <- as_bm_list("\U0001f42d\U0001f432\U0001f435", font = font)
  bm <- do.call(cbind, bml)
  print(bm, px = px_ascii)
}
}
