% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combining.R
\name{unifont_combining}
\alias{unifont_combining}
\title{Get combining character information}
\usage{
unifont_combining(upper = TRUE, csur = TRUE, unicode = FALSE)
}
\arguments{
\item{upper}{Include glyphs above the Unicode Basic Multilingual plane.}

\item{csur}{Include (Under-)Conscript Unicode Registry glyphs.}

\item{unicode}{Include combining glyphs assigned by the Unicode Consortium
(i.e. not ones in the Private Use Area like the CSUR ones).
By default \code{FALSE} since \code{\link[bittermelon:bm_compose]{bittermelon::bm_compose()}} can usually guess that
a Unicode Consortium assigned glyph is a combining glyph by
using \code{\link[Unicode:u_char_properties]{Unicode::u_char_property()}}.}
}
\value{
A character vector of Unicode code points
}
\description{
\code{unifont_combining()} returns a character vector of the code points
for all the "combining" characters in Unifont.
}
\examples{
  df <- unifont_combining()
  head(df)
}
\seealso{
Can be used with the \code{pua_combining} argument of \code{\link[bittermelon:bm_compose]{bittermelon::bm_compose()}}
and \code{\link[bittermelon:as_bm_bitmap]{bittermelon::as_bm_bitmap()}}.
}
