% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unifont.R
\name{unifont_version}
\alias{unifont_version}
\title{GNU Unifont version number}
\usage{
unifont_version()
}
\value{
The Unifont version number as a \code{\link[=numeric_version]{numeric_version()}} class.
}
\description{
The function \code{unifont_version()} returns the GNU Unifont version number
this package packed their \code{hex} files from.
}
\examples{
unifont_version()
}
