% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unifont.R
\name{unifont}
\alias{unifont}
\title{Load Unifont font}
\usage{
unifont(upper = TRUE, jp = FALSE, csur = TRUE, sample = FALSE)
}
\arguments{
\item{upper}{Include glyphs above the Unicode Basic Multilingual plane.}

\item{jp}{Use Japanese version of Chinese characters.}

\item{csur}{Include (Under-)Conscript Unicode Registry glyphs.}

\item{sample}{Add circle to "Combining" characters.}
}
\value{
A \code{\link[bittermelon:bm_font]{bittermelon::bm_font()}} object.
}
\description{
The function \code{unifont()} loads in several GNU Unifont hex files as a
single \code{\link[bittermelon:bm_font]{bittermelon::bm_font()}} object.
}
\examples{
if (require("bittermelon")) {
  # Note Unifont is a **big** font
  system.time(font <- unifont())
  length(font)
  object.size(font)

  # Mandarin Chinese
  as_bm_list("ï¼²å¾ˆæ£’ï¼", font = font) |> bm_call(cbind) |> print()

  # Emoji
  as_bm_list("ðŸ­ðŸ²ðŸµ", font = font) |> bm_call(cbind) |> print()
}
}
